refactor: format code
