+++ lib/core/domain/error/data_error.dart
