+++ lib/core/domain/error/error.dart
