+++ lib/domain/models/user.dart
@@ -3,9 +3,10 @@ class User {
   final String imageUrl;
   final String? name;
 
-   User({
+  User({
     this.id,
     String? imageUrl,
     this.name,
-  }) : imageUrl = imageUrl ?? 'https://eu.ui-avatars.com/api/?name=${name?.split('').join('+')}&size=250'  ;
+  }) : imageUrl = imageUrl ??
+            'https://eu.ui-avatars.com/api/?name=${name?.split('').join('+')}&size=250';
 }
