+++ lib/domain/use_cases/get_restaurants_use_case.dart
@@ -3,8 +3,6 @@ import 'package:restaurant_tour/core/domain/error/error.dart';
 import 'package:restaurant_tour/domain/models/restaurant.dart';
 import 'package:restaurant_tour/domain/repositories/restaurants_repository.dart';
 
-import '../../core/domain/error/data_error.dart';
-
 class GetRestaurantsUseCase {
   final BaseRestaurantsRepository _restaurantsRepository;
